bookstore-frontend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-Mail==0.9.1
Flask-SocketIO==5.1.1
python-dotenv==0.19.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
alembic==1.7.1
email-validator==1.1.3
Pillow==8.3.2
stripe==2.61.0
pytest==6.2.5
